Fixes how current task end time is updated
Refactors the name of the task list variable. Length of list is checked
each time a new task is started. If the list is not empty, the last task
on the list is given the current time as its end time. That task does
not need to be added to the list again.